support Date objects in json queries
Selecting a TIMESTAMP column in a plain select or plain join returns the
timestamp as a js Date objects like you'd expect it.

Using `selectAs` or `selectAsJsonAgg` generates SQL such that the timestamp
is embedded in a json object and returned as such to the node-postgres
driver which then does not apply any conversion :(

Fixed this by running optional in-place conversion functions based on the
schema definitions so that timestamps work as expected. Could be extended
to provide general serialization/deserialization to strings or database
blobs.