idea for better where expressions
After letting it sit in the back for a while, I do not like the current
stringy generic where - to noisy and it completely breaks the flow of the
chaining api.

But I still need a good enough where expression, maybe going for good old
lispy PRN will fix that itch for me?